Added maximum trauma
I Originally didn't want Trauma to have a limit, but I decided for the purpose of playability, it needs a limit.
Otherwise, a player could end up in a situation where their trauma is So high, that it isn't practical to reduce it via normal means.
But, that could happen so I decided to add a limit.